Make BasePagedList properties behave consistently in no-data cases
Refer to issue #137

The expressions to derive the values for 3 BasePagedList properties
have been changed to be more appropriate and inline with the others in
cases where the current page has no data.

HasPreviousPage when PageNumber > PageCount is now false instead of
true. IsFirstPage when PageCount = 0 is now false instead of true.
FirstItemOnPage and LastItemOnPage when PageCount = 0 or PageNumber >
PageCount are now both 0 instead of being calculated based on page
number and size.

Added test for the totalItemCount parameter in the BasePagedList
constructor being less than 0 - this is an error condition.

This work reverts commit [9cca40b5] (the fix for issue #115 (Empty
list causes exception) is now implemented in
X.PagedList.PagedListExtensions.ToPagedList) and [35ca2e5e].

Unit tests and documentation comments updated accordingly.