X.PagedList.Tests: PagedListTheories: Fix wrong source list
In ToListAsync_ForQueryable_With_TotalSetCount_Works, the point is to compare
the tested method ToPagedListAsync against the manual implementation with
Skip/Take and the two versions with and without superSetTotalCount against
each other.

However, instead of using the same source set for all three actions, one used
the result of one other which is obviously flawed.

Fix this.