X.PagedList.Mvc.Core: Enable nullable annotations
This enables nullable annotations in the projects and adjusts all types
accordingly. There are three unresolved warnings due to properties that are
assigned to non-nullable values when assigned, but have no default value.

This can be resolved by the "required" member attribute, but this is only
available with C# 11 (.NET 7). Keep the warnings for now.

Note that the Nullable definition is moved to Directory.Build.props now, since
we have it enabled in all core projects now.