Treewide: Provide specific namespaces for Extensions
At the moment, Extensions like ToPagedList and ToPagedListAsync are defined
in the same namespace as all the other types. This poses a few problems:

  * If one wants to use X.PagedList.EF, he will have name collisions between
    the EF version of ToPagedListAsync and the default async-over-sync version.
    While this can be resolved by choosing the X.PagedList.EF namespace in a
    "using", using any of the base types like IPagedList would still require
    to add the X.PagedList "using" again or use a full qualifier.

  * If somebody only wants the base types like IPagedList, but desires to
    implement Extensions on his own, he will draw back in the upstream
    Extensions by a "using" statement.

  * If somebody wants to hide the async-over-sync implementations from
    inexperienced developers, but needs the rest of the package, there is no
    option so far.

All of this can be resolved by putting the different Extensions in their
dedicated namespaces. This way one can choose whether he only wants the base
classes and interfaces or whether different types of Extensions are desired.
Choosing between them is as simple as putting the appropriate "using" on top
of the file. This allows to switch between the EF and async-over-sync
implementations as it does allow to disable some of them.

The new namespaces are:

  X.PagedList.Extensions        The sync implementations

  X.PagedList.AsyncExtensions   The fake async (async-over-sync) implementations

  X.PagedList.EF                The proper async EF Core implementations
                                (this namespace is unchanged)

This will require users of earlier package versions to add them too, of course,
but this is trivial.