X.PagedList: PagedList: Require non-null value for superset
There is not really a point to creating a PagedList without a superset.

If somebody actually desires to have an empty one, he should provide an empty
list.

This is a breaking change. If somebody actually created a PagedList with "null"
for whatever reason, this will now throw an Exception.