Treewide: Provide specific namespace for Extensions (again)
As discussed in [1], using a separate namespace for Extensions has several
advantages. Unfortunately, the dedicated namespace for the PagedListExtensions
has been removed during the package name rearrangements and not added back
after.

This commit adds it back and essentially reverts the namespace scheme to
where it was in v10.0.3, like already done for the rest of the code.

To indicate this more prominently, the PagedListExtensions.cs class is moved
into an "Extensions" subfolder, so the namespace now also matches the folder
structure.

From the released packages, this affects X.PagedList and X.PagedList.Mvc.Core.

[1] 6ae094e4bd45 ("Treewide: Provide specific namespaces for Extensions")

Fixes: #290